+++ app/src/main/java/com/itclimb/mylittlelemoncomposablelab/database/Database.kt
@@ -1,4 +1,4 @@
-package com.itclimb.mylittlelemoncomposablelab
+package com.itclimb.mylittlelemoncomposablelab.database
 
 
 import androidx.lifecycle.LiveData
