+++ app/src/main/java/com/itclimb/mylittlelemoncomposablelab/ui/Elements.kt
@@ -1,4 +1,4 @@
-package com.itclimb.mylittlelemoncomposablelab
+package com.itclimb.mylittlelemoncomposablelab.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
@@ -7,6 +7,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.itclimb.mylittlelemoncomposablelab.R
 
 
 @Composable
